requirements.txt
# Python dependencies for the Autonomous Agent
litellm
python-dotenv